w1-Assignment/requirements.txt
openai
transformers
anthropic
google-generativeai
streamlit
langdetect
json
datetime